requirements.txt
cookiecutter
GitPython